+++ src/NodesTranslator.ts
@@ -1,4 +1,6 @@
 import { XMutationObserver } from './lib/XMutationObserver';
+import { isInViewport } from './utils/isInViewport';
+import { nodeExplore } from './utils/nodeExplore';
 import { configureTranslatableNodePredicate } from './utils/nodes';
 
 interface NodeData {
@@ -27,46 +29,6 @@ interface NodeData {
 	 * Priority to translate node. The bigger the faster will translate
 	 */
 	priority: number;
-}
-
-/**
- * @param handler if return `false`, loop will stop
- */
-const nodeExplore = (
-	inputNode: Node,
-	nodeFilter: number,
-	includeSelf: boolean,
-	handler: (value: Node) => void | boolean,
-) => {
-	const walk = document.createTreeWalker(inputNode, nodeFilter, null);
-	let node = includeSelf ? walk.currentNode : walk.nextNode();
-	while (node) {
-		if (handler(node) === false) {
-			return;
-		}
-		node = walk.nextNode();
-	}
-};
-
-/**
- * Check visibility of element in viewport
- */
-export function isInViewport(element: Element, threshold = 0) {
-	const { top, left, bottom, right, height, width } = element.getBoundingClientRect();
-	const overflows = {
-		top,
-		left,
-		bottom: (window.innerHeight || document.documentElement.clientHeight) - bottom,
-		right: (window.innerWidth || document.documentElement.clientWidth) - right,
-	};
-
-	if (overflows.top + height * threshold < 0) return false;
-	if (overflows.bottom + height * threshold < 0) return false;
-
-	if (overflows.left + width * threshold < 0) return false;
-	if (overflows.right + width * threshold < 0) return false;
-
-	return true;
 }
 
 type TranslatorInterface = (text: string, priority: number) => Promise<string>;
